requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2.1
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
